Remove the git commit option
I wanted to do this for a while now.
As far as I know, the git commit option was not used. Anyway, it is
probably easy to emulate by just telling the program to run the
necessary commands (I can't say for sure if that's all it will take,
because of the problem of having to ignore the .git directory')

This change also made the code much cleaner and less criptic.