Add timer control to the status icon
I needed to create a Timer class to keep track of time.

This means we now have two related gobject timeouts: one for the timer
itself to decrement the time_left, and another one for the status icon
to update it's label.

The UserInterface object need to know the Timer object so it can control
it's state (pause, start, etc)